+++ app/src/main/java/com/android/t4tek/data/repository/UserRepositoryImpl.kt
@@ -1,19 +1,18 @@
 package com.android.t4tek.data.repository
 
 import android.util.Log
-import com.android.t4tek.data.api.ApiHelper
+import com.android.t4tek.data.remote.ApiHelper
 import com.android.t4tek.data.local.AppDatabase
 import com.android.t4tek.data.local.entity.User
+import com.android.t4tek.domain.BaseRepository
 import com.android.t4tek.domain.repository.UserRepository
-import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 import java.util.Random
-import javax.inject.Singleton
 
 class UserRepositoryImpl @Inject constructor(
     private val db: AppDatabase,
     private val apiHelper: ApiHelper
-) : UserRepository {
+) : UserRepository, BaseRepository() {
     override fun showLog() {
         Log.i("hanhmh1203", "showLog")
         insertRandomRecord()
