+++ app/src/main/java/com/android/t4tek/domain/BaseRepository.kt
@@ -1,4 +1,4 @@
 package com.android.t4tek.domain
 
-class BaseRepository {
+open class BaseRepository {
 }
